requirements.txt
crewai>=0.28.0
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
streamlit>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
exa-py>=0.6.0
duckduckgo-search>=4.1.0
